feat: implement a weekly/monthly calendar/scheduler view
