+++ apps/web/package.json
@@ -10,6 +10,7 @@
     "clean": "rm -rf .next"
   },
   "dependencies": {
+    "@bitnoi.se/react-scheduler": "^0.1.4",
     "@clerk/nextjs": "^4.26.2",
     "@edgestore/react": "^0.1.4",
     "@edgestore/server": "^0.1.4",
@@ -27,6 +28,7 @@
     "@trpc/server": "^10.43.6",
     "@vercel/analytics": "^1.1.1",
     "@vercel/edge": "^1.1.0",
+    "dayjs": "^1.11.10",
     "framer-motion": "^10.16.4",
     "lucide-react": "^0.292.0",
     "next": "^14.0.1",
