+++ apps/web/src/app/(authedLayout)/calendar/month/page.tsx
@@ -1,0 +1,5 @@
+import { CalendarGrid } from "@/calendar/components/CalendarGrid";
+
+export default function MontlyViewPage() {
+  return <CalendarGrid />;
+}
